% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vector_control.R
\name{add_irs_history}
\alias{add_irs_history}
\title{Add IRS History as Vertical Lines}
\usage{
add_irs_history(district_name, clr = "#4686FBFF", Yr0 = 2015)
}
\arguments{
\item{district_name}{a district name}

\item{clr}{a color for the labels}

\item{Yr0}{the start year}
}
\value{
invisible()
}
\description{
Add IRS History as Vertical Lines
}
