% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vector_control.R
\name{add_itn_history}
\alias{add_itn_history}
\title{Add itn History as Vertical Lines}
\usage{
add_itn_history(district_name, clr = "#E4460AFF", Yr0 = 2015)
}
\arguments{
\item{district_name}{a district name}

\item{clr}{a color for the labels}

\item{Yr0}{the start year}
}
\value{
invisible()
}
\description{
Add itn History as Vertical Lines
}
