% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{district_adjacency}
\alias{district_adjacency}
\title{District population data}
\format{
An object of class \code{data.frame} with 146 rows and 146 columns.
}
\usage{
district_adjacency
}
\description{
District population data
}
\keyword{datasets}
