% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{district_dir}
\alias{district_dir}
\title{District Directories}
\format{
An object of class \code{data.frame} with 146 rows and 7 columns.
}
\usage{
district_dir
}
\description{
District Directories
}
\keyword{datasets}
