% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sse_ts_function.R
\name{fit_ts_function}
\alias{fit_ts_function}
\title{Fit the amplitude for a time series}
\usage{
fit_ts_function(ts_data, times)
}
\arguments{
\item{data}{the PR observed}

\item{model}{an \code{xds} model}

\item{Fpar}{parameters from \code{makepar_F_sin}}
}
\value{
an \code{xds} object
}
\description{
For a time series \eqn{X,} compute the
phase, the time of the year when there is a peak \code{dog}
}
