% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sse_ts_function.R
\name{fit_ts_season}
\alias{fit_ts_season}
\title{Fit the amplitude for a time series}
\usage{
fit_ts_season(ts_data, times, Tpar)
}
\arguments{
\item{Tpar}{parameters for the interannual pattern}

\item{data}{the PR observed}

\item{model}{an \code{xds} model}
}
\value{
an \code{xds} object
}
\description{
For a time series \eqn{X,} compute the
phase, the time of the year when there is a peak \code{dog}
}
