% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sse_ts_function.R
\name{funky}
\alias{funky}
\title{Make a function of time}
\usage{
funky(avg, Spar, Tpar)
}
\arguments{
\item{avg}{the average}

\item{Spar}{parameters for the seasonal component}

\item{Tpar}{parameters for the interannual component}
}
\value{
a function
}
\description{
Return a function with the given
average, with seasonality and interannual variation
}
