% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfpr.R
\name{get_district_pfpr_i}
\alias{get_district_pfpr_i}
\title{Get a District PfPR, by Index}
\usage{
get_district_pfpr_i(i)
}
\arguments{
\item{district}{a district name}
}
\value{
a data frame
}
\description{
Get a District PfPR, by Index
}
