% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vector_control.R
\name{get_irs_jdates}
\alias{get_irs_jdates}
\title{Get Julian dates for IRS spray rounds}
\usage{
get_irs_jdates(district_name, Yr0 = 2015)
}
\arguments{
\item{district_name}{a district name}

\item{Yr0}{the start year}
}
\value{
a vector
}
\description{
Get Julian dates for IRS spray rounds
}
