% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/months.R
\name{get_month}
\alias{get_month}
\title{Get the Month Name(s)}
\usage{
get_month(i)
}
\arguments{
\item{i}{Month as a number}
}
\value{
Three letter abbreviation for month
}
\description{
Get the Month Name(s)
}
\examples{
get_month(2)
}
