% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_neighbors}
\alias{get_neighbors}
\title{Get Neighboring Districts}
\usage{
get_neighbors(ix)
}
\arguments{
\item{ix}{the index of a district}
}
\value{
a list of neighbors and their indices
}
\description{
Get Neighboring Districts
}
