% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfpr.R
\name{get_one_district}
\alias{get_one_district}
\title{Get a District PfPR, by Index}
\usage{
get_one_district(i)
}
\arguments{
\item{i}{district index}
}
\value{
a data frame
}
\description{
Get a District PfPR, by Index
}
