% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{half_pipe}
\alias{half_pipe}
\title{Reconstruction Pipeline}
\usage{
half_pipe(
  i,
  base_model_name,
  wd_name,
  impute_ty = "mean",
  trust_ty = "unmodified"
)
}
\arguments{
\item{i}{the index of a district to pull from}

\item{base_model_name}{the stem name of a model}

\item{wd_name}{the working directory name}

\item{impute_ty}{a text string to dispatch \code{impute_baseline_ty}}

\item{trust_ty}{a text string to dispatch \code{get_trust_ty}}
}
\value{
a
}
\description{
Run the baseline reconstruction pipeline for the
\eqn{i^{th}}
}
