% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shapes.R
\name{label_region}
\alias{label_region}
\title{Plot a Region with District Labels}
\usage{
label_region(region, xshift = c(), yshift = c())
}
\arguments{
\item{region}{the region name}

\item{xshift}{district specific x values to shift the label}

\item{yshift}{district specific y values to shift the label}
}
\value{
invisible()
}
\description{
Plot a Region with District Labels
}
