% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{make_bednet_history}
\alias{make_bednet_history}
\title{Make Bed Net Rounds Setup Object for a District}
\usage{
make_bednet_history(dname)
}
\arguments{
\item{dname}{the district name}
}
\value{
a bednet multi-round setup object
}
\description{
Make Bed Net Rounds Setup Object for a District
}
