% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{make_bednet_rounds}
\alias{make_bednet_rounds}
\title{Make Bed Net Rounds Setup Object for a District}
\usage{
make_bednet_rounds(dname, t_end = 3352)
}
\arguments{
\item{dname}{the district name}

\item{t_end}{a cutoff for late events}
}
\value{
a bednet multi-round setup object
}
\description{
Make Bed Net Rounds Setup Object for a District
}
