% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shapes.R
\name{make_district_icon}
\alias{make_district_icon}
\title{Plot a district in red within its region}
\usage{
make_district_icon(district)
}
\arguments{
\item{district}{the district name}
}
\value{
invisible()
}
\description{
Plot a district in red within its region
}
