% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{make_irs_history}
\alias{make_irs_history}
\title{Make IRS Rounds Setup Object for a District}
\usage{
make_irs_history(dname)
}
\arguments{
\item{dname}{the district name}
}
\value{
an IRS multi-round setup object
}
\description{
Make IRS Rounds Setup Object for a District
}
