% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{make_irs_rounds}
\alias{make_irs_rounds}
\title{Make IRS Rounds Setup Object for a District}
\usage{
make_irs_rounds(dname, t_end = 3352)
}
\arguments{
\item{dname}{the district name}

\item{t_end}{a cutoff for late events}
}
\value{
an IRS multi-round setup object
}
\description{
Make IRS Rounds Setup Object for a District
}
