% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pfpr_by_district}
\alias{pfpr_by_district}
\title{District population data}
\format{
An object of class \code{data.frame} with 16790 rows and 7 columns.
}
\usage{
pfpr_by_district
}
\description{
District population data
}
\keyword{datasets}
