% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vector_control.R
\name{plot_itn_history}
\alias{plot_itn_history}
\title{Create a Plot of ITN Mass Distributions for a District}
\usage{
plot_itn_history(
  district_name,
  start = 2012,
  end = 2026,
  add = FALSE,
  clr = "#E4460AFF"
)
}
\arguments{
\item{district_name}{a district name}

\item{start}{the first year to plot}

\item{end}{the last year to plot}

\item{add}{if TRUE, add to an existing plot}

\item{clr}{a color for the labels}
}
\value{
invisible()
}
\description{
Create a Plot of ITN Mass Distributions for a District
}
