% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profile.R
\name{profile_plot_i}
\alias{profile_plot_i}
\title{Profile a Fitted Model from a List}
\usage{
profile_plot_i(i, model_list, data_list, Yr0 = 2015)
}
\arguments{
\item{i}{index}

\item{model_list}{a list of models, as xds objects}

\item{data_list}{a list of data sets}

\item{Yr0}{a reference year}
}
\value{
invisible()
}
\description{
Profile a Fitted Model from a List
}
