% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profile.R
\name{profile_pr}
\alias{profile_pr}
\title{Plot Predicted and Observed PR}
\usage{
profile_pr(model, data, Yr0 = 2015)
}
\arguments{
\item{model}{a model; an xds object}

\item{data}{a data set}

\item{Yr0}{a reference year}
}
\value{
invisible()
}
\description{
Plot Predicted and Observed PR
}
