% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profile.R
\name{profile_residual}
\alias{profile_residual}
\title{Plot the residuals}
\usage{
profile_residual(model, data, Yr0 = 2015)
}
\arguments{
\item{model}{a model; an xds object}

\item{data}{a data set}

\item{Yr0}{a reference year}
}
\value{
invisible()
}
\description{
Plot the residuals
}
