% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profile.R
\name{profile_temporal}
\alias{profile_temporal}
\title{Plot Seasonal & Interannual Patterns}
\usage{
profile_temporal(model, Yr0 = 2015)
}
\arguments{
\item{model}{a model; an xds object}

\item{Yr0}{a reference year}
}
\value{
invisible()
}
\description{
Plot Seasonal & Interannual Patterns
}
