% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_fitting.R
\name{realized_history_ix}
\alias{realized_history_ix}
\title{Fit the Realized History}
\usage{
realized_history_ix(ix, base_model, model_name = "sip_eir")
}
\arguments{
\item{ix}{an index for a district}

\item{base_model}{a \strong{\code{ramp.xds}} model object with scaling}

\item{model_name}{a string, the model name}
}
\value{
the filename
}
\description{
Fit the Realized History
}
