% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{region_dir}
\alias{region_dir}
\title{District Directories}
\format{
An object of class \code{data.frame} with 15 rows and 4 columns.
}
\usage{
region_dir
}
\description{
District Directories
}
\keyword{datasets}
