% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_fitting.R
\name{setup_bednet_eval}
\alias{setup_bednet_eval}
\title{Setup Bednet Evaluation}
\usage{
setup_bednet_eval(xds_obj, round, N = 2, buf = 1000)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{round}{the round to evaluate}

\item{N}{the number of spline points}

\item{buf}{the buffer around the event}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Setup Bednet Evaluation
}
