% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shapes.R
\name{shade_district}
\alias{shade_district}
\title{Plot a district in red within its region}
\usage{
shade_district(district)
}
\arguments{
\item{district}{the district name}
}
\value{
invisible()
}
\description{
Plot a district in red within its region
}
