% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shapes.R
\name{shade_region}
\alias{shade_region}
\title{Plot Uganda Shading a Region}
\usage{
shade_region(region)
}
\arguments{
\item{region}{the region name}
}
\value{
invisible()
}
\description{
Plot Uganda Shading a Region
}
