% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_fitting.R
\name{short_fit_bednet_rounds}
\alias{short_fit_bednet_rounds}
\title{Evaluate the Bed Net Rounds}
\usage{
short_fit_bednet_rounds(ix, model_name = "sip_eir")
}
\arguments{
\item{ix}{an index for a district}

\item{model_name}{a string, the model name}
}
\value{
a string
}
\description{
Evaluate the Bed Net Rounds
}
