% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sse_ts_function.R
\name{sse_ts_function}
\alias{sse_ts_function}
\title{Given data, compute GoF for a seasonal function}
\usage{
sse_ts_function(ts_data, times, func)
}
\arguments{
\item{times}{the times of the observations}

\item{func}{the model}

\item{data}{the PR observed}
}
\value{
a list with the mean peak and the values
}
\description{
For a time series c(\code{times},\code{data}),
and a model, compute the sum of squared errors
}
