% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shapes.R
\name{uga_districts_plot}
\alias{uga_districts_plot}
\title{Standardized Plot for Uganda Districts}
\usage{
uga_districts_plot(clrs = "#E6E6E6", ttl = "Default Uganda by District")
}
\arguments{
\item{clrs}{The color(s) for the district}

\item{ttl}{The Plot Title}
}
\value{
invisible()
}
\description{
Standardized Plot for Uganda Districts
}
