% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uga_irs}
\alias{uga_irs}
\title{Uganda Mass ITN Distribution Data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 316 rows and 19 columns.
}
\usage{
uga_irs
}
\description{
Uganda Mass ITN Distribution Data
}
\keyword{datasets}
