% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uga_itn}
\alias{uga_itn}
\title{Uganda Mass ITN Distribution Data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 584 rows and 23 columns.
}
\usage{
uga_itn
}
\description{
Uganda Mass ITN Distribution Data
}
\keyword{datasets}
