% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uga_water_shp}
\alias{uga_water_shp}
\title{District population data}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 217 rows and 7 columns.
}
\usage{
uga_water_shp
}
\description{
District population data
}
\keyword{datasets}
