% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shapes.R
\name{outline_region}
\alias{outline_region}
\title{Plot a region outline}
\usage{
outline_region(region, add = TRUE)
}
\arguments{
\item{region}{the region name}

\item{add}{replot if add=FALSE}
}
\value{
invisible()
}
\description{
Plot a region outline
}
